refactor(uniffi): extract fire-uniffi-user and collapse FireAppCore to a pure facade
Phase 3f of the uniffi multi-namespace split (final domain extraction).

Rust:
- Add rust/crates/fire-uniffi-user with
  setup_scaffolding!("fire_uniffi_user"). Hosts FireUserHandle plus the
  user-profile / summary / badge / follow / user-action / invite record
  types. All 10 user-domain methods (fetch_user_profile, fetch_user_summary,
  fetch_following, fetch_followers, follow_user, unfollow_user,
  fetch_badge_detail, fetch_user_actions, fetch_pending_invites,
  create_invite_link) move here.
- InviteCreateRequestState migrates from fire-uniffi-topics to
  fire-uniffi-user with the create_invite_link method. The type was only
  colocated with topics because the legacy state_topic_detail.rs module had
  grouped every composer/upload/invite input together; semantically it
  belongs with the invite API.
- Delete fire-uniffi/src/handle.rs and fire-uniffi/src/state_user.rs.
  fire-uniffi/src/lib.rs now matches the plan's terminal shape: the three
  top-level #[uniffi::export] free functions (plain_text_from_html,
  preview_text_from_html, monogram_for_username) plus the FireAppCore facade
  that composes one Arc<FireXxxHandle> per domain and exposes a getter per
  domain. FireAppCore no longer holds inner / panic_state directly; every
  call site that needs them goes through Arc<SharedFireCore> via the domain
  handles.

iOS native:
- FireSessionStore.swift routes user / profile / follow / badge / invite
  calls through core.user().

Build plumbing:
- native/ios-app/project.yml preBuildScript outputs gain
  fire_uniffi_user.swift. Regenerated project.pbxproj.

cargo test --workspace stays green.
