+++ native/ios-app/project.yml
@@ -115,6 +115,7 @@ targets:
           - $(SRCROOT)/Generated/FireUniFfi/fire_uniffi_session.swift
           - $(SRCROOT)/Generated/FireUniFfi/fire_uniffi_topics.swift
           - $(SRCROOT)/Generated/FireUniFfi/fire_uniffi_types.swift
+          - $(SRCROOT)/Generated/FireUniFfi/fire_uniffi_user.swift
           - $(SRCROOT)/Generated/fire_uniffiFFI/fire_uniffiFFI.h
           - $(SRCROOT)/Generated/fire_uniffiFFI/module.modulemap
           - $(SRCROOT)/Generated/lib/$(PLATFORM_NAME)/libfire_uniffi.a
